* moved top_three_factors to it's own file to facilitate refactor
